requirements.txt
Flask == 2.2.5
Flask-API == 3.1
Flask-Cors == 4.0.0
Flask-SQLAlchemy == 3.1.1
python-dotenv == 1.0.1
bcrypt == 4.1.2
Werkzeug == 3.0.1
apscheduler == 3.10.4
build == 1.2.1
black-with-tabs == 22.10.0